Rooms fixed and player's animations remade
All rooms fixed, but walls and enemies still aren't working. Note that
the room now are in a "ObjectLayer", which is the only one to interact
with the player.
Except for the player's death animation, its every other has been
remade. No enemy animation has beed remade by me, as of now.